Moved arithmetic operations into their respective *Functions object
