+++ src/main/scala/Truncation.scala
@@ -251,4 +251,53 @@ object TruncationFunctions {
     Truncation(leaves.toVector.sorted(leftRightOrd))
 
   def rootTruncation() : Truncation = Truncation(Vector(rootLabel))
+
+  private object DropHead extends Enumeration {
+    type DropHead = Value
+    val L1, L2, Both = Value
+  }
+  import DropHead._
+
+  /**
+    * The leaves of the tree resulting from `RPUnion(t1, t2)`
+    */
+  def rpUnion(t1: Truncation, t2: Truncation): Truncation = {
+
+    def getDrop(n1: NodeLabel, n2: NodeLabel): (Vector[NodeLabel], DropHead) = {
+      if (n1 == n2)
+        (Vector(n1), Both)
+      else if (isAncestorOf(n2, n1))
+        ((n1 +: n1.sibling +: n1.ancestors.take(n1.depth - n2.depth - 1).map(_.sibling).toVector).sorted(leftRightOrd), Both)
+      else if (isAncestorOf(n1, n2))
+        ((n2 +: n2.sibling +: n2.ancestors.take(n2.depth - n1.depth - 1).map(_.sibling).toVector).sorted(leftRightOrd), Both)
+      else if (isLeftOf(n1, n2))
+        (Vector(n1), L1)
+      else
+        (Vector(n2), L2)
+    }
+
+    @scala.annotation.tailrec
+    def unionHelper(
+      acc: Vector[NodeLabel],
+      l1: Vector[NodeLabel], 
+      l2: Vector[NodeLabel]
+    ): Vector[NodeLabel] = {
+      if (l1.isEmpty)
+        acc ++ l2
+      else if (l2.isEmpty)
+        acc ++ l1
+      else {
+        val n1 = l1.head
+        val n2 = l2.head
+        val (next, drop) = getDrop(n1, n2)
+        unionHelper(
+          acc ++ next.filter( !acc.contains(_) ), 
+          if (drop != L2) l1.tail else l1,
+          if (drop != L1) l2.tail else l2
+        )
+      }
+    }
+
+    Truncation(unionHelper(Vector.empty, t1.leaves, t2.leaves))
+  }
 }
